Refactored int-float conversions for io. Closes #43

+++ librosa/feature.py
@@ -836,7 +836,8 @@ def sync(data, frames, aggregate=None):
 
     (dimension, n_frames) = data.shape
 
-    frames = np.unique(np.concatenate(([0], frames, [n_frames]))).astype(int)
+    frames = np.unique(np.concatenate(([0], frames, [n_frames])))
+    frames = frames.astype(int)
 
     if min(frames) < 0:
         raise ValueError('Negative frame index.')
